[Refactor] /domain/preuser 패키지 /domain/user 패키지로 리네이밍
