+++ src/main/java/com/example/ssccwebbe/domain/user/dto/PreUserRequestDto.java
@@ -1,4 +1,4 @@
-package com.example.ssccwebbe.domain.preuser.dto;
+package com.example.ssccwebbe.domain.user.dto;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
