+++ src/main/java/com/example/ssccwebbe/domain/user/entity/SocialProviderType.java
@@ -1,4 +1,4 @@
-package com.example.ssccwebbe.domain.preuser.entity;
+package com.example.ssccwebbe.domain.user.entity;
 
 import lombok.Getter;
