+++ src/main/java/com/example/ssccwebbe/domain/user/repository/PreUserRepository.java
@@ -1,10 +1,10 @@
-package com.example.ssccwebbe.domain.preuser.repository;
+package com.example.ssccwebbe.domain.user.repository;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.example.ssccwebbe.domain.preuser.entity.PreUserEntity;
+import com.example.ssccwebbe.domain.user.entity.PreUserEntity;
 
 public interface PreUserRepository extends JpaRepository<PreUserEntity, Long> {
