+++ src/main/java/com/example/ssccwebbe/domain/user/service/PreUserService.java
@@ -1,10 +1,10 @@
-package com.example.ssccwebbe.domain.preuser.service;
+package com.example.ssccwebbe.domain.user.service;
 
 import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
 import org.springframework.security.oauth2.client.userinfo.OAuth2UserService;
 import org.springframework.security.oauth2.core.user.OAuth2User;
 
-import com.example.ssccwebbe.domain.preuser.dto.PreUserResponseDto;
+import com.example.ssccwebbe.domain.user.dto.PreUserResponseDto;
 
 public interface PreUserService extends OAuth2UserService<OAuth2UserRequest, OAuth2User> {
     PreUserResponseDto readPreUser();
