+++ src/main/java/com/example/ssccwebbe/domain/user/service/PreUserServiceImpl.java
@@ -1,4 +1,4 @@
-package com.example.ssccwebbe.domain.preuser.service;
+package com.example.ssccwebbe.domain.user.service;
 
 import java.util.List;
 import java.util.Map;
@@ -15,13 +15,13 @@ import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.example.ssccwebbe.domain.preuser.code.PreUserErrorCode;
-import com.example.ssccwebbe.domain.preuser.dto.CustomOAuth2PreUser;
-import com.example.ssccwebbe.domain.preuser.dto.PreUserRequestDto;
-import com.example.ssccwebbe.domain.preuser.dto.PreUserResponseDto;
-import com.example.ssccwebbe.domain.preuser.entity.PreUserEntity;
-import com.example.ssccwebbe.domain.preuser.entity.SocialProviderType;
-import com.example.ssccwebbe.domain.preuser.repository.PreUserRepository;
+import com.example.ssccwebbe.domain.user.code.PreUserErrorCode;
+import com.example.ssccwebbe.domain.user.dto.CustomOAuth2PreUser;
+import com.example.ssccwebbe.domain.user.dto.PreUserRequestDto;
+import com.example.ssccwebbe.domain.user.dto.PreUserResponseDto;
+import com.example.ssccwebbe.domain.user.entity.PreUserEntity;
+import com.example.ssccwebbe.domain.user.entity.SocialProviderType;
+import com.example.ssccwebbe.domain.user.repository.PreUserRepository;
 import com.example.ssccwebbe.global.apipayload.exception.GeneralException;
 import com.example.ssccwebbe.global.security.UserRoleType;
