+++ src/main/java/com/example/ssccwebbe/global/security/config/ScheduleConfig.java
@@ -7,7 +7,7 @@ import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.example.ssccwebbe.domain.preuser.repository.PreUserRefreshRepository;
+import com.example.ssccwebbe.domain.user.repository.PreUserRefreshRepository;
 
 import lombok.RequiredArgsConstructor;
