+++ src/test/java/com/example/ssccwebbe/domain/admin/service/ApplyFormAdminServiceTest.java
@@ -20,7 +20,7 @@ import com.example.ssccwebbe.domain.applyform.entity.ApplyFormStatus;
 import com.example.ssccwebbe.domain.applyform.entity.CodingExp;
 import com.example.ssccwebbe.domain.applyform.repository.ApplyFormInterviewTimeRepository;
 import com.example.ssccwebbe.domain.applyform.repository.ApplyFormRepository;
-import com.example.ssccwebbe.domain.preuser.entity.PreUserEntity;
+import com.example.ssccwebbe.domain.user.entity.PreUserEntity;
 
 @ExtendWith(MockitoExtension.class)
 class ApplyFormAdminServiceTest {
