+++ src/test/java/com/example/ssccwebbe/domain/user/entity/PreUserRefreshEntityTest.java
@@ -1,4 +1,4 @@
-package com.example.ssccwebbe.domain.preuser.entity;
+package com.example.ssccwebbe.domain.user.entity;
 
 import static org.assertj.core.api.Assertions.assertThat;
